order 취소버튼 활성화, 반복요청시 Alert 호출

+++ src/views/shippingpoint/shippingpoint.js
@@ -2,22 +2,21 @@
  *
  * Author: 박상준
  * date : 2022-05-31
- * Todo:
  */
 
 import * as Api from '/api.js';
 import store from '../cart/store.js';
 import { addCommas, searchAddressByDaumPost } from '/useful-functions.js';
 
-// 바로구매확인
-const url = new URL(window.location.href);
-const id = url.searchParams.get('id');
-
 // localStorage data
-const storedItem = store
+const storedItem =
+location.search?
+  [ JSON.parse( sessionStorage.getItem('product'))]
+  :
+  store
   .getLocalStorage()
   .filter((item) => item.cart === 'checked');
-
+console.log(storedItem);
 //user Inputs
 const receiverNameInput = document.getElementById('receiverNameInput');
 const receiverPhoneNumberInput = document.getElementById(
@@ -69,11 +68,6 @@ async function insertAddressToAddrInputs() {
 }
 
 function insertOrderSummary() {
-  //바로구매
-  if(id !== null){
-        
-  }
-
   if (!storedItem || storedItem.length < 1) return;
   console.log(storedItem);
   let amount = 0;
@@ -100,7 +94,6 @@ function insertOrderSummary() {
 //회원정보 수정 진행
 async function sendOrderInfoByPost(e) {
   e.preventDefault();
-
   const receiverName = receiverNameInput.value;
   const receiverPhoneNumber = receiverPhoneNumberInput.value;
   const postalCode = postalCodeInput.value;
@@ -165,11 +158,15 @@ async function sendOrderInfoByPost(e) {
     messageTo: requestComment,
     products,
   };
-  console.log(order);
+
   try {
     const result = await Api.post('/api/orders', order);
     if (result) {
-      localStorage.clear();
+      if(Product){
+        sessionStorage.setItem('Product','');
+      }else{
+        localStorage.clear();
+      }
       window.location.href="/complete";
       return alert('성공적으로 주문했습니다.');
     }
